redoes equality, again
Previously, we had === for equality that typechecks and =:= for equality
that normalizes.

However, today's discussion at the Scala meeting made me realize that
=== is a bad name for a typechecking equality (thanks @sjrd) + I found
a real use case for === as an equality that doesn't typecheck, doesn't
normalize, but just goes ahead and compares (possibly incomplete) attributes.
Therefore, now we have === (dumb), =:= (typechecking) and =~= (normalizing).

Moreover, while trying to make scalahost work, I realized that we need
custom equality for Denotation (because Denotation.prefix is a Tree),
Typing (because Typing.tpe is a Tree) and Expansion (because Expansion.desugaring is a Tree).

This prompted a generalization of the previously implemented Equality and
Equivalence helper objects to handle not just trees, but everything else
that can contain them.

Ideally, we'd have this as a type class, but that would require a significant
effort to write down (or generate) all the cases for the dozens of our AST nodes,
and time is a luxury that I don't have at this particular stage of development.
Therefore I'm just putting an ugly Any-based implementation in place for now
and marking this as a TODO item.
