+++ docs/hosts.md
@@ -14,7 +14,7 @@ Hosts are, however, required to create instances of scala.meta trees to be retur
 
   1. Trees are fully immutable in the sense that they: a) don't contain any observationally mutable fields, b) aren't supposed contain any references to anything that might be mutable. This means that after creation trees can never change and as such can be easily reasoned about. In order to "modify" an existing tree, one is supposed to use either `copy` or one of the tree transformers (e.g. `Tree.transform` or other, more advanced, functionality of the TQL framework). Both of these approaches create a copy of an original tree with fields changed appropriately.
 
-  NB! Since trees can't be mutated in-place, reference equality (`Tree.eq` and `Tree.ne`) should not be used to work with trees. On a related note, standard equality (`Tree.==`) is as helpful as reference equality, because it's deliberately implemented to redirect to `Tree.eq`. Structural comparisons should be performed with quasiquotes, and semantic comparisons should be done with `Tree.===` or `Tree.=:=`.
+  NB! Since trees can't be mutated in-place, reference equality (`Tree.eq` and `Tree.ne`) should not be used to work with trees. On a related note, standard equality (`Tree.==`) is as helpful as reference equality, because it's deliberately implemented to redirect to `Tree.eq`. Structural comparisons should be performed with quasiquotes, and semantic comparisons should be done with `Tree.=:=` or `Tree.=~=`.
 
   1. Trees strongly emphasize safety by construction in the sense that they disallow creation of syntactically invalid language constructs (e.g. it's impossible to create trees that represent classes without primary constructors or applications of types to terms). As much validation as possible is pushed to compile time via precise types of tree fields, and some leftovers whose validation was too heavyweight to encode in types are verified at runtime. For you as a host implementor this means that you'll have to be really precise in arguments that you provide to our tree creation facilities. Don't worry though. Even if you make a mistake, it'll manifest itself by a compilation error or a crash immediately upon creation of the malformed tree, not by corrupting some internal state and crashing mysteriously afterwards.
